Add an update loop API so plugins can hook in their main loops
In case they need to busywait or deal with a message loop. Heavy lifting should be done using a custom solution.

+++ src/Driver/NodeView.h
@@ -4,7 +4,7 @@
 #include <vector>
 
 struct NodeView {
-	enum class NodeType : uint64_t {
+	enum class NodeType : uint32_t {
 		Unknown = 0,
 		Haptic = 1,
 		Led = 2,
@@ -23,7 +23,7 @@ struct NodeView {
 		Data DisplayData;
 
 	};
-	uint64_t region;
+	uint32_t region;
 
 	std::vector<SingleNode> nodes;
 };
